+++ src/About.ts
@@ -54,7 +54,7 @@ export class About implements IClassAbout {
 
   public async apiDefaults(path: string) {
     if (!path)
-      throw new Error(`aboutApiDefaults: "path" parameter is required`);
+      throw new Error(`about.apiDefaults: "path" parameter is required`);
     return await this.repoClient
       .Get(`about/api/default/${path}`)
       .then((res) => res.data);