+++ src/Stream.ts
@@ -1,8 +1,9 @@
 import { QlikRepositoryClient } from "./main";
 import { UpdateCommonProperties } from "./util/UpdateCommonProps";
 import { GetCommonProperties } from "./util/GetCommonProps";
+import { URLBuild } from "./util/generic";
 
-import { IHttpStatus, IEntityRemove, IOwner } from "./types/interfaces";
+import { IEntityRemove, IOwner, ISelection } from "./types/interfaces";
 import { ICustomPropertyCondensed } from "./CustomProperty";
 import { ITagCondensed } from "./Tag";
 
@@ -43,6 +44,8 @@ export interface IClassStream {
   getFilter(filter: string): Promise<IStream[]>;
   create(name: IStreamCreate): Promise<IStream>;
   remove(id: string): Promise<IEntityRemove>;
+  removeFilter(filter: string): Promise<IEntityRemove[]>;
+  select(filter?: string): Promise<ISelection>;
   update(arg: IStreamUpdate): Promise<IStream>;
 }
 
@@ -53,7 +56,7 @@ export class Stream implements IClassStream {
   }
 
   public async get(id: string) {
-    if (!id) throw new Error(`steamGet: "id" parameter is required`);
+    if (!id) throw new Error(`steam.get: "id" parameter is required`);
     return await this.repoClient
       .Get(`stream/${id}`)
       .then((res) => res.data as IStream);
@@ -67,7 +70,7 @@ export class Stream implements IClassStream {
 
   public async getFilter(filter: string) {
     if (!filter)
-      throw new Error(`streamGetFilter: "filter" parameter is required`);
+      throw new Error(`stream.getFilter: "filter" parameter is required`);
 
     return await this.repoClient
       .Get(`stream/full?filter=(${encodeURIComponent(filter)})`)
@@ -76,7 +79,7 @@ export class Stream implements IClassStream {
 
   public async create(arg: IStreamCreate) {
     if (!arg.name)
-      throw new Error(`streamCreate: "path" parameter is required`);
+      throw new Error(`stream.create: "path" parameter is required`);
 
     let getCommonProps = new GetCommonProperties(
       this,
@@ -96,14 +99,40 @@ export class Stream implements IClassStream {
   }
 
   public async remove(id: string) {
-    if (!id) throw new Error(`streamRemove: "id" parameter is required`);
+    if (!id) throw new Error(`stream.remove: "id" parameter is required`);
     return await this.repoClient.Delete(`stream/${id}`).then((res) => {
       return { id, status: res.status } as IEntityRemove;
     });
   }
 
+  public async removeFilter(filter: string) {
+    if (!filter)
+      throw new Error(`stream.removeFilter: "filter" parameter is required`);
+
+    const tags = await this.getFilter(filter).then((t: IStream[]) => {
+      if (t.length == 0)
+        throw new Error(`stream.removeFilter: filter query return 0 items`);
+
+      return t;
+    });
+    return await Promise.all<IEntityRemove>(
+      tags.map((tag: IStream) => {
+        return this.remove(tag.id);
+      })
+    );
+  }
+
+  public async select(filter?: string) {
+    const urlBuild = new URLBuild(`selection/stream`);
+    urlBuild.addParam("filter", filter);
+
+    return await this.repoClient
+      .Post(urlBuild.getUrl(), {})
+      .then((res) => res.data as ISelection);
+  }
+
   public async update(arg: IStreamUpdate) {
-    if (!arg.id) throw new Error(`streamUpdate: "id" parameter is required`);
+    if (!arg.id) throw new Error(`stream.update: "id" parameter is required`);
 
     let stream = await this.get(arg.id);